Added button component.
For this it was required that the MouseEvent from buildin SaxionApp be overridden by a custom one. SaxionApp does not allow mouse movement tracking by default.
Our framework now supports this. It only tracks mouse when inside the game window, and automatically scales this value to the render resolution instead of the window resolution.
Also added a method in Painter class to draw a rectangle outline. wanted this to confirm location of button bounds.
Also added a toString override for Vector2
All places that implemented the SaxionApp MouseEvent class have been changed to implement the new MouseEvent class